updating login and balance
