+++ lib/api_service/balance_api.dart
@@ -1,0 +1,20 @@
+import 'package:http/http.dart' as http;
+import 'dart:convert';
+
+Future<String> fetchBalance(String username, String password) async {
+  var data = {
+    "username": username,
+    "password": password,
+    "action": "balance", // Use a different action to fetch the balance
+  };
+
+  final response = await http
+      .post(Uri.parse("https://portal.fastsmsnigeria.com/api/?"), body: data);
+
+  if (response.statusCode == 200) {
+    var responseData = jsonDecode(response.body);
+    return responseData['balance'].toString();
+  } else {
+    return "Error fetching balance";
+  }
+}
