+++ lib/screens/mobile_extractor_screen.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:nbsms/api_service/api.dart';
 import 'package:nbsms/constant/constant_colors.dart';
 import 'package:nbsms/constant/constant_fonts.dart';
 import 'package:nbsms/constant/constant_mediaquery.dart';
@@ -10,6 +11,7 @@ import 'package:nbsms/widgets/body_singlescroll_widget.dart';
 import 'package:nbsms/widgets/drawer_widget.dart';
 import 'package:nbsms/widgets/page_title.dart';
 import 'package:nbsms/widgets/submit_button.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 
 class MobileExScreen extends StatefulWidget {
   const MobileExScreen({super.key});
@@ -22,6 +24,26 @@ class _MobileExScreenState extends State<MobileExScreen> {
   bool isExtract = false;
   checkExtract() {
     isExtract = true;
+  }
+
+  String balance = "";
+  Future<void> _fetchBalance() async {
+    SharedPreferences prefs = await SharedPreferences.getInstance();
+    String username = prefs.getString('username') ?? '';
+    String password = prefs.getString('password') ?? '';
+
+    String fetchedBalance = await fetchBalance(
+        username, password); // Call the method from api_service.dart
+    setState(() {
+      balance =
+          fetchedBalance; // Update the balance variable with the fetched value
+    });
+  }
+
+  @override
+  void initState() {
+    super.initState();
+    _fetchBalance();
   }
 
   @override
@@ -70,7 +92,7 @@ class _MobileExScreenState extends State<MobileExScreen> {
             style: TextStyle(fontFamily: centurygothic, fontSize: 16.0),
             children: <TextSpan>[
               TextSpan(text: '₦', style: TextStyle(fontFamily: roboto)),
-              const TextSpan(text: '31,000.00'),
+              TextSpan(text: balance),
             ],
           ),
         ),
